# University_Club_Project_YMH220-main/University_Club_Project_YMH220-main/mock_api/openapi.yaml
openapi: 3.0.3
info:
  title: UniClub API Specification
  description: API for the UniClub University Club Management System. This spec ensures a seamless API First development integration between frontend and backend teams.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
    description: Local Mock Server
  - url: http://127.0.0.1:8000
    description: Local FastAPI Backend Server (Under Development)
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
              example:
                email: student@university.edu
                password: securepassword123
                full_name: John Doe
      responses:
        '201':
          description: Created
          content:
            application/json:
              example:
                id: 99
                email: student@university.edu
                full_name: John Doe
                role: student
  /api/v1/auth/login:
    post:
      summary: Login for access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: student@university.edu
                password: securepassword123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              example:
                access_token: mock-jwt-token-123
                token_type: bearer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT